Move cluster writers in separate file
